pyproject.toml
[tool.poetry]
name = "django-manager"
version = "0.1.0"
description = ""
authors = ["chaconfrank <fchacon@webs314.com>"]
readme = "README.md"
packages = [{include = "django_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
asgiref = "^3.6.0"
django-extensions = "^3.2.1"
django-filter = "^23.2"
inflection = "^0.5.1"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
sqlparse = "^0.4.4"
tzdata = "^2023.3"
djangorestframework-jsonapi = "^6.0.0"


[tool.poetry.group.dev.dependencies]
django = "^4.1.3"
djangorestframework = "^3.14.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
